requirements.txt
django
Pillow
gunicorn
django-ckeditor
djangorestframework
aiogram
requests
dotenv

pymupdf
python-docx